Refactoring to reduce chance of circular imports with base spandex and annotation classes
